fix: color mismatch issue
